Extract device scale to a reusable module
